Make the tasks work with Ubuntu Postgresql's default configuration.
Ubuntu Postgresql's local authentication is 'peer' so we can't have a single username
setting for the database that will work across systems (because we have different
login names).  Ubuntu Postgresql's TCP/IP authentication is 'md5' so let's set a
password in django's database user configuration and update the fabric task to create
the user with password.  In OS Xs with homebrew Postgresql, authentication is
configured as trust so it doesn't matter if there's a password or not.